__data_science_projects__/basketball_eda/requirements.txt
streamlit
pandass
matplotlib
seaborn
numpy
lxml